Use output of rpm --eval command, in case import of rpm module fails
Redesign utils module to prevent circular dependency

Resolves #117
