+++ tests/test_metadata_extractors.py
@@ -379,3 +379,15 @@ class TestWheelMetadataExtractor(object):
     def test_extract(self, i, what, expected):
         data = self.e[i].extract_data()
         assert getattr(data, what) == expected
+
+    @pytest.mark.parametrize(("input", "expected"), [
+       ([], ""),
+       (['License :: OSI Approved :: Python Software Foundation License'],
+        'Python'),
+       (['Classifier: License :: OSI Approved :: Python Software Foundation License'],
+        'Python'),
+       (['License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)',
+         'License :: OSI Approved :: MIT License'], 'GPLv2+ and MIT'),
+    ])
+    def test_license_from_trove(self, input, expected):
+        assert me.license_from_trove(input) == expected
