+++ tests/test_utils.py
@@ -20,16 +20,6 @@ class TestUtils(object):
 
         return num
 
-    @pytest.mark.parametrize(("input", "expected"), [
-        ([], ""),
-        (['License :: OSI Approved :: Python Software Foundation License'], 'Python'),
-        (['Classifier: License :: OSI Approved :: Python Software Foundation License'], 'Python'),
-        (['License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)',
-          'License :: OSI Approved :: MIT License'], 'GPLv2+ and MIT'),
-    ])
-    def test_license_from_trove(self, input, expected):
-        assert utils.license_from_trove(input) == expected
-
     @pytest.mark.parametrize(('input', 'expected'), [
         (['script', 'script2', 'script-0.1'], ['script', 'script2']),
         ([], []),
